requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.0.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
Pillow>=10.0.0
